backend/requirements.txt
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
pillow==10.0.0
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
gdown==4.7.1